// types/project.ts
export interface Project {
  id: string
  name: string
  slug: string
  uri : string
  description: string
  imageUrl: string
  category: string
  location: string
  visitorCount: number
  storeCount: number
  rating: number
  createdAt: string
  adminName: string
  adminAvatar: string
  floors: number
}